sunnyside/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.12.10
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.1
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
dill==0.3.5.1
Django==4.1.1
django-cleanup==6.0.0
django-next-prev==1.1.0
docutils==0.19
exceptiongroup==1.0.0rc9
frozenlist==1.3.1
future==0.18.2
geocoder==1.38.1
geoip2==4.6.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
ip2geotools==0.1.6
IP2Location==8.8.1
isort==5.10.1
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.9.3
lazy-object-proxy==1.7.1
lxml==4.9.1
maxminddb==2.2.0
mccabe==0.7.0
more-itertools==8.14.0
multidict==6.0.2
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pip-review==1.2.0
pkginfo==1.8.3
platformdirs==2.5.2
psycopg==3.1.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
pylint==2.15.3
pyparsing==3.0.9
pyquery==1.4.3
PySocks==1.7.1
ratelim==0.1.6
readme-renderer==37.2
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.6.0
SecretStorage==3.3.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
twine==4.0.1
typed-ast==1.5.4
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
zipp==3.8.1
django-on-heroku==1.1.2